+++ tailwind.config.js
@@ -1,7 +1,36 @@
 const plugin = require('tailwindcss/plugin')
 
+const borderRadius = Object.fromEntries(
+  Object.values(require('./tokens/dist/js/rounded').rounded)
+    // Filter out aliases here because they are only included for backward compatibility with Scss
+    .filter(({ original}) => original.value.includes('{') === false)
+    .map(({ attributes, value }) => [attributes.type, value])
+)
+
+const boxShadow = Object.fromEntries(
+  Object.values(require('./tokens/dist/js/boxShadow')['box-shadow'])
+    // Filter out aliases here because they are only included for backward compatibility with Scss
+    .filter(({ original}) => original.value.includes('{') === false)
+    .map(({ attributes, value }) => [attributes.type, value])
+)
+
+const screens = Object.fromEntries(
+  Object.values(require('./tokens/dist/js/breakpoints').breakpoints).map(
+    ({ attributes, value }) => [attributes.type, value]
+  )
+)
+
+const zIndex = Object.fromEntries(
+  Object.values(require('./tokens/dist/js/zIndex')['z-index']).map(
+    ({ attributes, value }) => [attributes.type, value]
+  )
+)
+
 module.exports = {
-  content: ['./components/**/*.{js,vue}', './directives/**/*.js'],
+  content: [
+    './src/components/**/*.{js,vue}',
+    './src/directives/**/*.js'
+  ],
   important: true, // Utilities should always win https://sebastiandedeyne.com/why-we-use-important-with-tailwind/
   plugins: [
     plugin(function({ addUtilities }) {
@@ -24,6 +53,10 @@ module.exports = {
     }
   ],
   theme: {
+    screens,
+    borderRadius,
+    boxShadow,
+    zIndex,
     extend: {
       flexShrink: {
         2: '2'